backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
transformers==4.30.0
torch==2.1.0
faiss-cpu==1.7.3
pydantic==2.2.0
pymupdf==1.20.2  # or pdfplumber==0.5.28 if you prefer


